se agrega ValidacionAeropuertoDataEnum para que devuelva el error especifico segun se necesite al intentar crear un aeropuerto. A su vez se crean dos test para probar las verificaciones.

+++ src/main/java/ar/com/ada/api/aladas/services/VueloService.java
@@ -73,8 +73,8 @@ public class VueloService {
     }
 
     public boolean validarAeropuertoOrigenDiffDestino(Vuelo vuelo) {
-     
-        return vuelo.getAeropuertoDestino() != vuelo.getAeropuertoOrigen();
+
+        return vuelo.getAeropuertoDestino().intValue() != vuelo.getAeropuertoOrigen().intValue();
 
     }
